Make VSCodeOperations partial
